requirements.txt
requests>=2.31.0
web3>=6.15.1
yfinance>=0.2.36
pandas>=2.1.4
pyarrow>=15.0.0
python-binance>=1.0.19
python-telegram-bot>=20.7
python-dotenv>=1.0.0
matplotlib>=3.8.0
mplfinance>=0.12.10
